Lego WireLess Protocoll implementation
