+++ LegoBTLE/LegoWP/DNS_MSG_HUB_ALERT.py
@@ -27,8 +27,7 @@
 from dataclasses import dataclass, field
 
 from LegoBTLE.LegoWP.common_message_header import COMMON_MESSAGE_HEADER
-from LegoBTLE.LegoWP.hub_alert import HUB_ALERT, HUB_ALERT_OPERATION
-from LegoBTLE.LegoWP.m_type import M_TYPE
+from LegoBTLE.LegoWP.types import HUB_ALERT, HUB_ALERT_OPERATION, M_TYPE
 
 
 @dataclass
